refactor(knowledge-base): split ChunksMetadataFilter into focused modules
ChunksMetadataFilter was holding data fetching, derived option lists,
validation rules, popover orchestration, and the combobox UI all in one
file. Split along single-responsibility lines so each piece changes for
its own reason:

  - MetadataCombobox.tsx — presentational Popover+Command combobox with
    custom-entry fallback. Generic over options/testId; reusable.
  - hooks/useChunksMetadataFilter.ts — wraps useGetKbMetadataKeys and
    derives availableKeys / valueSuggestions / flags / refetch.
  - metadataFilterValidation.ts — pure KEY_PATTERN + validateMetadataFilter
    returning { ok, key, value } | { ok: false, error }. Testable in
    isolation; no DOM.
  - ChunksMetadataFilter.tsx — orchestration only: outer popover, form
    state, submit. Composes the three above.

Adds focused unit tests for the validator (regex + happy path + empty
+ uppercase). Existing integration test for the filter itself is left
untouched.
